Began work to make notejam work with AWS RDS
Notejam is a very simple monolithic app. It uses sqlite
to run a simple database on the same server as the app.
My goal is to instead have Notejam use an external db,
specifically Amazon RDS. The changes made in this commit
make that possible and can easily be modified to have
Notejam work with any external database endpoint.

This is kind of a messy initial commit. It adds a bunch
of terraform files as well as a few other config scripts.
The notejam schema is also there and a placeholder shell
script. Goal is to be able to spin up a working notejam
environment with one command using terraform, docker,
Amazon's container service, and RDS. Then we will
integrate with CodeDeploy and have a complete sample
CI/CD pipeline.