Several updates to create fully automated builds
This is still very much a work in progress...

We add our base Dockerfile. This will create an
image we can use for the notejam app.

This commit includes an initial buildspec.yml file
for AWS ECS. It also has several updates to the
terraform files and some new ones. We have a file to
automate creation of the AWS container repo. There
is also a script for creating an iam policy.

This commit also adds a really hacky performance
test skeleton in the form of a curl command and
Python script.