Schema update and setup.sh script improvements
Ran into a few errors after doing more testing on
the app. Turns out that AUTO_INCREMENT needed to
be set on all three tables.

Also, trying to get setup.sh to do more work. I
want the user to be able to run that and have it
deal with the docker image build and push to the
Amazon container repo. Eventually I may make this
script handle the terraform tasks too.