Added almost working serverless config using Zappa
Followed this guide:
https://www.agiliq.com/blog/2019/01/complete-serverless-django/

Ran into a bunch of compatibility issues with Zappa that all seem to stem from the fact that Django 1.6.5 is too old.
Fixed most of them by adding extra requirements and forcing versions but ran into one again when running the DB creation script
provided in the guide. (https://stackoverflow.com/a/43785773)

Not really sure why Zappa is calling setup. The older version of Zappa I picked seemed to be OK with Django 1.6.5 in previous steps.

Not spending any more time to investigate.

Virtualenv is pretty huge and will not be commited but should be easy to recreate via the requirements.txt